+++ src/main/java/com/ayakacraft/carpetayakaaddition/CarpetAyakaAddition.java
@@ -21,27 +21,18 @@
 package com.ayakacraft.carpetayakaaddition;
 
 import carpet.CarpetServer;
+import com.ayakacraft.carpetayakaaddition.utils.AyakaLanguage;
 import com.ayakacraft.carpetayakaaddition.utils.mods.ModUtils;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.google.gson.reflect.TypeToken;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.loader.api.ModContainer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.lang.reflect.Type;
-import java.nio.charset.StandardCharsets;
-import java.util.*;
+import java.util.Optional;
 
 public final class CarpetAyakaAddition implements ModInitializer {
-
-    private static final Type MAP_TYPE = new TypeToken<Map<String, String>>() {
-    }.getType();
-
-    private static final Map<String, Map<String, String>> translations = new HashMap<>(2);
 
     public static final Gson GSON = new GsonBuilder().setPrettyPrinting().setLenient().create();
 
@@ -65,38 +56,10 @@ public final class CarpetAyakaAddition implements ModInitializer {
         LOGGER = LogManager.getLogger(MOD_NAME);
     }
 
-    public static Map<String, String> getTranslations(String lang) {
-        if (translations.containsKey(lang)) {
-            return translations.get(lang);
-        }
-        Map<String, String> translation = Collections.emptyMap();
-        final InputStream   langStream  = CarpetAyakaServer.class.getClassLoader().getResourceAsStream(String.format("assets/carpet-ayaka-addition/lang/%s.json", lang));
-        if (langStream != null) { // otherwise we don't have that language
-            final String jsonData;
-            try {
-                byte[] data = new byte[langStream.available()];
-                int    i    = langStream.read(data);
-                if (i != data.length) {
-                    data = Arrays.copyOf(data, i);
-                }
-                jsonData = new String(data, StandardCharsets.UTF_8);
-                langStream.close();
-
-                Map<String, String> map = GSON.fromJson(jsonData, MAP_TYPE);
-                if (map != null) {
-                    translation = map;
-                    translations.put(lang, translation);
-                }
-            } catch (final IOException ignored) {
-            }
-        }
-
-        return translation;
-    }
-
     @Override
     public void onInitialize() {
         LOGGER.debug("Initializing {} Version {}", CarpetAyakaAddition.MOD_NAME, CarpetAyakaAddition.MOD_VERSION);
+        AyakaLanguage.loadLanguages();
         CarpetServer.manageExtension(CarpetAyakaServer.INSTANCE);
     }